requirements.txt
accelerate
bitsandbytes
diffusers==0.31.0
transformers>=4.45.2
huggingface_hub
hf_transfer>=0.1.8
peft>=0.12.0
decord>=0.6.0
wandb
pandas
torchao>=0.5.0
sentencepiece>=0.2.0
imageio-ffmpeg>=0.5.1
numpy>=1.26.4
deepspeed
opencv-python
imageio
easydict